test(protocol): isolate index-name + extract shared test-db helpers
Code-review fixes for 5b670c77:
- afterEach now restores setConfigIndexName so a mid-test throw can't
  leak the catalog scope to the next test.
- initTestDatabase and seedCollection move to specs/helpers/db-helpers.ts
  (renamed from test-db.ts to avoid the test-*.ts gitignore pattern).
  specs/mcp.test.ts retains its inline copy pending a follow-up.
